[VS-7293] lookup query service introduced
