+++ restitution-app/Startup.cs
@@ -42,6 +42,7 @@ namespace Gov.Cscp.VictimServices.Public
 
             services.AddSingleton<IHttpContextAccessor, HttpContextAccessor>();
             services.AddTransient<TokenHandler>();
+            services.AddScoped<LookupQueryService>();
 
             services.AddHttpClient<ICOASTAuthService, COASTAuthService>();
